backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
Pillow==11.0.0
python-decouple==3.8
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
gunicorn==23.0.0
stripe>=5.0
